Rename `spark_lib` -> `spark_dialect`
