+++ cohortextractor/query_engines/spark_dialect.py
