+++ tests/lib/databases.py
@@ -14,7 +14,9 @@ MSSQL_SETUP_DIR = Path(__file__).parents[1].absolute() / "support/mssql"
 
 # Register our modified PyHive SQLAlchemy dialect
 registry.register(
-    "spark.pyhive.opensafely", "cohortextractor.query_engines.spark_lib", "SparkDialect"
+    "spark.pyhive.opensafely",
+    "cohortextractor.query_engines.spark_dialect",
+    "SparkDialect",
 )
